# Speed up Uniform Duration sampling.