These changes speed up the process of sampling Durations from a uniform distribution primarily by avoiding two expensive operations:
- Creating a Duration (now only done once per sample call)
- Addition of Durations (now never done)

I also added some simple benchmarks to measure these changes.
Before:
```
test distr_uniform_duration_large   ... bench:      13,007 ns/iter (+/- 189) = 1230 MB/s
test distr_uniform_duration_largest ... bench:       9,263 ns/iter (+/- 136) = 1727 MB/s
test distr_uniform_duration_one     ... bench:      12,346 ns/iter (+/- 353) = 1295 MB/s
test distr_uniform_duration_variety ... bench:      12,371 ns/iter (+/- 130) = 1293 MB/s
```
After:
```
test distr_uniform_duration_large   ... bench:      12,523 ns/iter (+/- 41) = 1277 MB/s
test distr_uniform_duration_largest ... bench:       8,395 ns/iter (+/- 102) = 1905 MB/s
test distr_uniform_duration_one     ... bench:       9,907 ns/iter (+/- 422) = 1615 MB/s
test distr_uniform_duration_variety ... bench:      10,504 ns/iter (+/- 90) = 1523 MB/s
```
To really avoid any regressions, I also used a more comprehensive benchmark. 
The code for those benchmarks (rust to bench and python to graph) can be found [here](https://gist.github.com/Pazzaz/91aa3c475ead8096daae2f0d7af7574b).
<details> <summary>Comprehensive Benchmark Results</summary>


In these heatmaps, the x axis is the upper bound and the y axis is the lower bound. 800 different Durations were used, ranging from 0s to 18446744074.7095512s, in order.
The z-value is the average number of nanoseconds it took to sample from the corresponding distribution. 
Before:

![screenshot from 2018-08-04 21-16-25](https://user-images.githubusercontent.com/16006944/43688562-7d46ecb8-98eb-11e8-84a8-23927f4ca933.png)

After:
![screenshot from 2018-08-04 21-16-36](https://user-images.githubusercontent.com/16006944/43688563-7d5d8f40-98eb-11e8-9f51-fd350d260551.png)
Percentage of improvement:
![screenshot from 2018-08-05 21-58-40](https://user-images.githubusercontent.com/16006944/43689424-cf66fb46-98fa-11e8-8f61-6d7b501ff8bd.png)
Zoomed in:
![screenshot from 2018-08-04 21-15-40](https://user-images.githubusercontent.com/16006944/43688561-7d2d6fa4-98eb-11e8-8e90-803f30bcce9b.png)
So about a 10-20% improvement and the values outside of that range seem to be just noise.
</details>

Edit: changed a call to `Duration::from_nanos` to `Duration::new(nanos / 1_000_000_000, (nanos % 1_000_000_000) as u32)` for compatibility with older rust versions. Shouldn't make much of a difference but the benchmark can be rerun.