@dhardy 2018-08-10
By the way: your extended benchmarks show a significant performance regression for one set of bounds (I don't know what these are). I'm not keen on accepting this as-is due to the potential for very poor performance on certain bounds (but if we *did*, then there should definitely be a benchmark highlighting the problem).