@Pazzaz 2018-08-14
I hadn't really considered the worst case. I think it should have a much better worst case performance now. I reduced the rejection zone by using an offset of `-low_n` in the Large case and made Small applicable in more cases. I've spent quite some time trying to optimize this further but I think it's good enough now.

Current bench-difference
old:
```
running 5 tests
test distr_uniform_duration_edge    ... bench:      13,522 ns/iter (+/- 59) = 1183 MB/s
test distr_uniform_duration_large   ... bench:      12,853 ns/iter (+/- 100) = 1244 MB/s
test distr_uniform_duration_largest ... bench:       8,945 ns/iter (+/- 17) = 1788 MB/s
test distr_uniform_duration_one     ... bench:      12,555 ns/iter (+/- 142) = 1274 MB/s
test distr_uniform_duration_variety ... bench:      12,555 ns/iter (+/- 185) = 1274 MB/s
```
new:
```
running 5 tests
test distr_uniform_duration_edge    ... bench:       5,641 ns/iter (+/- 69) = 2836 MB/s
test distr_uniform_duration_large   ... bench:      12,517 ns/iter (+/- 33) = 1278 MB/s
test distr_uniform_duration_largest ... bench:       8,352 ns/iter (+/- 32) = 1915 MB/s
test distr_uniform_duration_one     ... bench:       9,942 ns/iter (+/- 36) = 1609 MB/s
test distr_uniform_duration_variety ... bench:      12,085 ns/iter (+/- 185) = 1323 MB/s
```
One thing that is a little weird right now is that I've set the offset used in the Large case as a field in the struct instead of as a field on the enum itself. This is simply because I found it to improve the benchmarks.

> your extended benchmarks show a significant performance regression for one set of bounds 

I think that's just noise caused by my computer prioritizing something else.